requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.2.0
langchain-text-splitters==0.2.0
openpyxl==3.1.2
pdfplumber==0.10.0
ebooklib==0.18
python-docx==1.1.0
python-pptx==0.6.23
pypdf2==3.0.1
markdown==3.5.2
beautifulsoup4==4.12.2
pandas==2.2.0
